+++ packages/core/src/Popover/modifiers.js
@@ -1,3 +1,4 @@
+import { getBaseModifiers } from '../Popper/modifiers.js'
 import { ARROW_SIZE } from './Arrow.js'
 
 const BORDER_RADIUS = 4
@@ -11,35 +12,51 @@ const computeArrowPadding = () => {
     return Math.ceil(padding)
 }
 
-export const offset = {
-    name: 'offset',
-    options: {
-        offset: [0, ARROW_SIZE],
-    },
+const hideArrowWhenDisplaced = ({ state }) => {
+    const halfArrow = ARROW_SIZE / 2
+    const displacement = state.modifiersData.preventOverflow
+    const referenceRect = state.rects.reference
+    const shouldHideArrow =
+        Math.abs(displacement.x) >= referenceRect.width + halfArrow ||
+        Math.abs(displacement.y) >= referenceRect.height + halfArrow
+
+    if (typeof state.attributes.arrow !== 'object') {
+        state.attributes.arrow = {}
+    }
+
+    state.attributes.arrow['data-arrow-hidden'] = shouldHideArrow
+
+    return state
 }
 
-export const arrow = {
-    name: 'arrow',
-    options: {
-        padding: computeArrowPadding(),
-    },
-}
+export const combineModifiers = (arrow, arrowElement, resizeObservers) => {
+    const baseModifiers = getBaseModifiers(resizeObservers)
 
-export const hideArrowWhenDisplaced = {
-    name: 'hideArrowWhenDisplaced',
-    enabled: true,
-    phase: 'main',
-    fn: ({ state }) => {
-        const halfArrow = ARROW_SIZE / 2
-        const displacement = state.modifiersData.preventOverflow
-        const referenceRect = state.rects.reference
-        const shouldHideArrow =
-            Math.abs(displacement.x) >= referenceRect.width + halfArrow ||
-            Math.abs(displacement.y) >= referenceRect.height + halfArrow
+    if (!arrow) {
+        return baseModifiers
+    }
 
-        state.attributes.popper['data-popper-arrow-displaced'] = shouldHideArrow
-
-        return state
-    },
-    requires: ['preventOverflow'],
+    return [
+        ...baseModifiers,
+        {
+            name: 'offset',
+            options: {
+                offset: [0, ARROW_SIZE],
+            },
+        },
+        {
+            name: 'arrow',
+            options: {
+                padding: computeArrowPadding(),
+                element: arrowElement,
+            },
+        },
+        {
+            name: 'hideArrowWhenDisplaced',
+            enabled: true,
+            phase: 'main',
+            fn: hideArrowWhenDisplaced,
+            requires: ['preventOverflow'],
+        },
+    ]
 }