+++ packages/core/CHANGELOG.md
@@ -1,3 +1,24 @@
+## [4.21.1](https://github.com/dhis2/ui-core/compare/v4.21.0...v4.21.1) (2020-05-19)
+
+
+### Bug Fixes
+
+* prevent icon from shrinking when label text overflows ([ccd86bc](https://github.com/dhis2/ui-core/commit/ccd86bc269a1dc2a6121388c465f59a3b0705cbb))
+
+# [4.21.0](https://github.com/dhis2/ui-core/compare/v4.20.0...v4.21.0) (2020-05-19)
+
+
+### Features
+
+* add support for virtual elements in popper and popover ([3a06526](https://github.com/dhis2/ui-core/commit/3a065266f390501eb8cb606fe84f7ba81621c0dc))
+
+# [4.20.0](https://github.com/dhis2/ui-core/compare/v4.19.1...v4.20.0) (2020-05-18)
+
+
+### Features
+
+* let popper and popover accept react-refs or dom nodes directly ([df11f3c](https://github.com/dhis2/ui-core/commit/df11f3ca20d31094d36d3b7965d61364947f62ee))
+
 ## [4.19.1](https://github.com/dhis2/ui-core/compare/v4.19.0...v4.19.1) (2020-04-30)